pyproject.toml
[project]
name = "aws-project-planning"
version = "0.1.0"
description = "AWS Project Planning Tool - Create SOWs, Architecture Diagrams, and Price Estimates"
authors = [
    {name = "AWS Project Planning Team", email = "example@example.com"},
]
dependencies = [
    "click>=8.1.0",
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "python-docx>=0.8.11",
    "pyyaml>=6.0",
    "jinja2>=3.1.2",
    "boto3>=1.26.0",
    "diagrams>=0.23.3",
    "graphviz>=0.20.1",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
aws-planner = "aws_project_planning.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --cov=aws_project_planning"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.black]
line-length = 100
target-version = ['py310'] 